requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-pymongo==2.3.0
flask-jwt-extended==4.5.2
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
bcrypt==4.0.1
gunicorn==21.2.0
